+++ irs-registry-client/src/main/java/org/eclipse/tractusx/irs/registryclient/central/DigitalTwinRegistryClientImpl.java
@@ -22,6 +22,8 @@
  ********************************************************************************/
 package org.eclipse.tractusx.irs.registryclient.central;
 
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
 import java.util.List;
 import java.util.Map;
 
@@ -29,9 +31,10 @@ import io.github.resilience4j.retry.annotation.Retry;
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.IdentifierKeyValuePair;
-import org.eclipse.tractusx.irs.data.StringMapper;
+import org.eclipse.tractusx.irs.registryclient.decentral.ShellQueryBody;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -42,7 +45,6 @@ import org.springframework.web.util.UriComponentsBuilder;
 public class DigitalTwinRegistryClientImpl implements DigitalTwinRegistryClient {
 
     private static final String PLACEHOLDER_AAS_IDENTIFIER = "aasIdentifier";
-    private static final String PLACEHOLDER_ASSET_IDS = "assetIds";
     private final RestTemplate restTemplate;
     private final String descriptorEndpoint;
     private final String shellLookupEndpoint;
@@ -56,17 +58,6 @@ public class DigitalTwinRegistryClientImpl implements DigitalTwinRegistryClient 
 
         ensureUrlContainsPlaceholders(descriptorEndpoint, "digitalTwinRegistry.descriptorEndpoint",
                 PLACEHOLDER_AAS_IDENTIFIER);
-        ensureUrlContainsPlaceholders(shellLookupEndpoint, "digitalTwinRegistry.shellLookupEndpoint",
-                PLACEHOLDER_ASSET_IDS);
-    }
-
-    private void ensureUrlContainsPlaceholders(final String bpdmUrl, final String configPath,
-            final String... placeholders) {
-        if (StringUtils.isNotBlank(bpdmUrl)) {
-            for (final var placeholder : placeholders) {
-                require(bpdmUrl, configPath, placeholder);
-            }
-        }
     }
 
     private static void require(final String bpdmUrl, final String configPath, final String placeholder) {
@@ -81,20 +72,36 @@ public class DigitalTwinRegistryClientImpl implements DigitalTwinRegistryClient 
         return "{" + placeholderIdType + "}";
     }
 
+    private static String encodeWithBase64(final String aasIdentifier) {
+        return Base64.getEncoder().encodeToString(aasIdentifier.getBytes(StandardCharsets.UTF_8));
+    }
+
+    private void ensureUrlContainsPlaceholders(final String bpdmUrl, final String configPath,
+            final String... placeholders) {
+        if (StringUtils.isNotBlank(bpdmUrl)) {
+            for (final var placeholder : placeholders) {
+                require(bpdmUrl, configPath, placeholder);
+            }
+        }
+    }
+
     @Override
     @Retry(name = "registry")
     public AssetAdministrationShellDescriptor getAssetAdministrationShellDescriptor(final String aasIdentifier) {
         final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(descriptorEndpoint);
-        final Map<String, String> values = Map.of(PLACEHOLDER_AAS_IDENTIFIER, aasIdentifier);
+        final Map<String, String> values = Map.of(PLACEHOLDER_AAS_IDENTIFIER, encodeWithBase64(aasIdentifier));
         return restTemplate.getForObject(uriBuilder.build(values), AssetAdministrationShellDescriptor.class);
     }
 
     @Override
     @Retry(name = "registry")
     public List<String> getAllAssetAdministrationShellIdsByAssetLink(final List<IdentifierKeyValuePair> assetIds) {
-        final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(shellLookupEndpoint);
-        final var values = Map.of(PLACEHOLDER_ASSET_IDS, StringMapper.mapToString(assetIds));
-        return restTemplate.exchange(uriBuilder.build(values), HttpMethod.GET, null,
+        final ShellQueryBody queryBody = ShellQueryBody.builder()
+                                                       .query(ShellQueryBody.ShellQuery.builder()
+                                                                                       .assetIds(assetIds)
+                                                                                       .build())
+                                                       .build();
+        return restTemplate.exchange(shellLookupEndpoint, HttpMethod.POST, new HttpEntity<>(queryBody),
                 new ParameterizedTypeReference<List<String>>() {
                 }).getBody();
     }