+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Integrity chain validation based on DataIntegrity aspect
 
+## [3.3.3] - 2023-08-11
+### Changed
+- IRS now calls the entire dataplane URL retrieved from the registry href instead of building it from the URL of the EDC token and the path
+
+### Fixed
+- Switched to POST for DTR lookup request
+- Added Base64 encoding to identifier for DTR shell-descriptor request 
+- Fixed an issue where IRS did not pass the BPN correctly for the ESS use-case
+
 ## [3.3.2] - 2023-07-31
 ### Fixed
 - BPN is now passed on correctly when traversing the item graph
@@ -300,7 +309,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Unresolved
 - **Select Aspects you need**  You are able to select the needed aspects for which you want to collect the correct endpoint information.
 
-[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/3.3.2...HEAD
+[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/3.3.3...HEAD
+[3.3.3]: https://github.com/eclipse-tractusx/item-relationship-service/compare/3.3.2...3.3.3
 [3.3.2]: https://github.com/eclipse-tractusx/item-relationship-service/compare/3.3.1...3.3.2
 [3.3.1]: https://github.com/eclipse-tractusx/item-relationship-service/compare/3.3.0...3.3.1
 [3.3.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/3.2.1...3.3.0